# Optimize catchpointdump utility