## Summary
* Do not load everything into memory but write directly in into a file
  This creates extremely high number of reallocation and CPU and RAM wasting.
* Add cmd options for downloading only (no loading), and for downloading
  from a single relay

# Perf data:
| param       | current | new      |
|-----------|--------|---------|
| RAM          | 5 GB    | 200 MB |
| Total time | 823 s   | 441 s    |

Downloading time went down from minutes to tens of seconds.